Move the rest of the logic
