+++ Content.Server/Wires/WiresSystem.cs
@@ -29,7 +29,7 @@ public sealed class WiresSystem : SharedWiresSystem
     [Dependency] private readonly SharedInteractionSystem _interactionSystem = default!;
     [Dependency] private readonly UserInterfaceSystem _uiSystem = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
-    [Dependency] private readonly Shared.Construction.ConstructionSystem _construction = default!;
+    [Dependency] private readonly ConstructionSystem _construction = default!;
 
     private static readonly ProtoId<ToolQualityPrototype> CuttingQuality = "Cutting";
     private static readonly ProtoId<ToolQualityPrototype> PulsingQuality = "Pulsing";
